+++ lms/services/lti_grading/_interface.py
@@ -1,0 +1,6 @@
+class LTIGradingClient:
+    def read_result(self, grading_id):
+        raise NotImplementedError()
+
+    def record_result(self, grading_id, score=None, pre_record_hook=None):
+        raise NotImplementedError()
